[refactor] move highlightSearchText to SearchBar.js
